Refactor Consumer and Supplier pages
