+++ src/features/amUI/maskinporten/DelegatedResourcesSection.tsx
@@ -1,0 +1,141 @@
+import { Button, DsAlert, DsHeading, DsParagraph, DsSkeleton } from '@altinn/altinn-components';
+import { MinusCircleIcon } from '@navikt/aksel-icons';
+import * as React from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { useFilteredResources } from '@/features/amUI/common/ResourceList/useFilteredResources';
+import { useIsMobileOrSmaller } from '@/resources/utils/screensizeUtils';
+import type {
+  ResourceDelegation,
+  ServiceResource,
+} from '@/rtk/features/singleRights/singleRightsApi';
+
+import { DelegationAction, EditModal } from '../common/DelegationModal/EditModal';
+import { ScopeList } from './ScopeList';
+import { getMaskinportenScopes } from './scopeUtils';
+import classes from './MaskinportenPage.module.css';
+
+interface DelegatedResourcesSectionProps {
+  resourcePermissions: ResourceDelegation[] | undefined;
+  isFetching: boolean;
+  hasError: boolean;
+  onRemove: (resource: ServiceResource) => void;
+  isResourceLoading: (resourceId: string) => boolean;
+  addNewResourceButton?: React.ReactNode;
+  editModalAvailableActions?: DelegationAction[];
+}
+
+export const DelegatedResourcesSection = ({
+  resourcePermissions,
+  isFetching,
+  hasError,
+  onRemove,
+  isResourceLoading,
+  addNewResourceButton,
+  editModalAvailableActions,
+}: DelegatedResourcesSectionProps) => {
+  const { t } = useTranslation();
+  const isMobile = useIsMobileOrSmaller();
+  const [search, setSearch] = React.useState('');
+  const [filterState, setFilterState] = React.useState<string[]>([]);
+  const [selectedResource, setSelectedResource] = React.useState<ServiceResource | null>(null);
+  const scopeModalRef = React.useRef<HTMLDialogElement>(null);
+
+  const delegatedResources = (resourcePermissions ?? [])
+    .map((delegation) => delegation.resource)
+    .filter((resource) => resource.identifier);
+
+  const { resources: filteredResources } = useFilteredResources<ServiceResource>({
+    resources: delegatedResources,
+    searchString: search,
+    serviceOwnerFilter: filterState,
+    getResourceName: (resource) => resource.title ?? '',
+    getOwnerName: (resource) => resource.resourceOwnerName ?? '',
+    getOwnerOrgCode: (resource) => resource.resourceOwnerOrgcode ?? '',
+    getDescription: (resource) => {
+      const scopes = getMaskinportenScopes(resource)
+        .map((ref) => ref.reference)
+        .join(' ');
+      return `${resource.description ?? ''} ${scopes}`.trim();
+    },
+  });
+
+  const serviceOwnerOptions = React.useMemo(() => {
+    const seen = new Map<string, { value: string; label: string }>();
+    for (const resource of delegatedResources) {
+      const code = resource.resourceOwnerOrgcode;
+      if (!code || seen.has(code)) continue;
+      seen.set(code, {
+        value: code,
+        label: resource.resourceOwnerName ?? code,
+      });
+    }
+    return Array.from(seen.values());
+  }, [delegatedResources]);
+
+  return (
+    <section className={classes.resourceSection}>
+      <DsHeading
+        level={2}
+        data-size='xs'
+      >
+        {t('maskinporten_page.delegated_resources_heading', { count: delegatedResources.length })}
+      </DsHeading>
+      {hasError ? (
+        <DsAlert data-color='danger'>
+          <DsParagraph>{t('maskinporten_page.delegated_resources_error')}</DsParagraph>
+        </DsAlert>
+      ) : isFetching && !resourcePermissions ? (
+        <DsSkeleton
+          width='100%'
+          height='2.5rem'
+        />
+      ) : (
+        <ScopeList
+          addNewResourceButton={addNewResourceButton}
+          resources={filteredResources}
+          search={search}
+          setSearch={setSearch}
+          filterState={filterState}
+          setFilterState={setFilterState}
+          serviceOwnerOptions={serviceOwnerOptions}
+          onSelect={(resource) => {
+            setSelectedResource(resource);
+            scopeModalRef.current?.showModal();
+          }}
+          renderControls={(resource) => {
+            const resourceLoading = isResourceLoading(resource.identifier);
+            return (
+              <Button
+                variant='tertiary'
+                size='sm'
+                loading={resourceLoading}
+                disabled={resourceLoading}
+                onClick={() => {
+                  onRemove(resource);
+                }}
+                aria-label={t('common.delete_poa_for', { poa_object: resource.title })}
+              >
+                <MinusCircleIcon aria-hidden='true' />
+                {!isMobile && t('common.delete_poa')}
+              </Button>
+            );
+          }}
+          emptyState={
+            <DsParagraph data-size='md'>
+              {delegatedResources.length === 0
+                ? t('maskinporten_page.no_delegated_resources')
+                : t('resource_list.no_resources_filtered', { searchTerm: search })}
+            </DsParagraph>
+          }
+        />
+      )}
+      <EditModal
+        ref={scopeModalRef}
+        maskinportenScope={selectedResource ?? undefined}
+        availableActions={editModalAvailableActions}
+        onClose={() => setSelectedResource(null)}
+      />
+    </section>
+  );
+};
