+++ src/features/amUI/maskinporten/SupplierPageContent.tsx
@@ -1,49 +1,32 @@
-import {
-  Button,
-  DsAlert,
-  DsHeading,
-  DsParagraph,
-  DsSkeleton,
-  formatDisplayName,
-  useSnackbar,
-} from '@altinn/altinn-components';
+import { Button, formatDisplayName, useSnackbar } from '@altinn/altinn-components';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
-import { MinusCircleIcon, TrashIcon } from '@navikt/aksel-icons';
+import { TrashIcon } from '@navikt/aksel-icons';
 
-import { useIsMobileOrSmaller } from '@/resources/utils/screensizeUtils';
 import {
   useGetMaskinportenSupplierResourcesQuery,
   useRemoveMaskinportenSupplierMutation,
   useRemoveMaskinportenSupplierResourceMutation,
 } from '@/rtk/features/maskinportenApi';
 import type { ServiceResource } from '@/rtk/features/singleRights/singleRightsApi';
-import { useFilteredResources } from '@/features/amUI/common/ResourceList/useFilteredResources';
 import { amUIPath } from '@/routes/paths';
 import { PartyType } from '@/rtk/features/userInfoApi';
 
-import { ScopeList } from './ScopeList';
-import { getMaskinportenScopes } from './scopeUtils';
 import { useMaskinportenResourceActions } from './hooks/useMaskinportenResourceActions';
-import classes from './MaskinportenPage.module.css';
-import { DelegationAction, EditModal } from '../common/DelegationModal/EditModal';
+import { DelegationAction } from '../common/DelegationModal/EditModal';
 import { DelegationModal, DelegationType } from '../common/DelegationModal/DelegationModal';
 import { PageContainer } from '../common/PageContainer/PageContainer';
 import { UserPageHeader } from '../common/UserPageHeader/UserPageHeader';
 import { usePartyRepresentation } from '../common/PartyRepresentationContext/PartyRepresentationContext';
 import { MaskinportenDeleteDialog } from './MaskinportenDeleteDialog';
+import { DelegatedResourcesSection } from './DelegatedResourcesSection';
 
 export const SupplierPageContent = () => {
   const { t } = useTranslation();
-  const isMobile = useIsMobileOrSmaller();
   const navigate = useNavigate();
   const { openSnackbar } = useSnackbar();
   const { fromParty, toParty } = usePartyRepresentation();
-  const [search, setSearch] = React.useState('');
-  const [filterState, setFilterState] = React.useState<string[]>([]);
-  const [selectedResource, setSelectedResource] = React.useState<ServiceResource | null>(null);
-  const scopeModalRef = React.useRef<HTMLDialogElement>(null);
   const deleteDialogRef = React.useRef<HTMLDialogElement>(null);
   const [removeSupplier, { isLoading: isRemovingSupplier }] =
     useRemoveMaskinportenSupplierMutation();
@@ -69,38 +52,6 @@ export const SupplierPageContent = () => {
         resource: resource.identifier,
       }).unwrap(),
   });
-
-  const delegatedResources = (resourcePermissions ?? [])
-    .map((delegation) => delegation.resource)
-    .filter((resource) => resource.identifier);
-
-  const { resources: filteredResources } = useFilteredResources<ServiceResource>({
-    resources: delegatedResources,
-    searchString: search,
-    serviceOwnerFilter: filterState,
-    getResourceName: (resource) => resource.title ?? '',
-    getOwnerName: (resource) => resource.resourceOwnerName ?? '',
-    getOwnerOrgCode: (resource) => resource.resourceOwnerOrgcode ?? '',
-    getDescription: (resource) => {
-      const scopes = getMaskinportenScopes(resource)
-        .map((ref) => ref.reference)
-        .join(' ');
-      return `${resource.description ?? ''} ${scopes}`.trim();
-    },
-  });
-
-  const serviceOwnerOptions = React.useMemo(() => {
-    const seen = new Map<string, { value: string; label: string }>();
-    for (const resource of delegatedResources) {
-      const code = resource.resourceOwnerOrgcode;
-      if (!code || seen.has(code)) continue;
-      seen.set(code, {
-        value: code,
-        label: resource.resourceOwnerName ?? code,
-      });
-    }
-    return Array.from(seen.values());
-  }, [delegatedResources]);
 
   const handleRemove = (resource: ServiceResource) =>
     remove(resource, {
@@ -159,82 +110,27 @@ export const SupplierPageContent = () => {
         direction='to'
         displayRoles={false}
       />
-      <section className={classes.resourceSection}>
-        <DsHeading
-          level={2}
-          data-size='xs'
-        >
-          {t('maskinporten_page.delegated_resources_heading', { count: delegatedResources.length })}
-        </DsHeading>
-        {resourcesError ? (
-          <DsAlert data-color='danger'>
-            <DsParagraph>{t('maskinporten_page.delegated_resources_error')}</DsParagraph>
-          </DsAlert>
-        ) : isFetching && !resourcePermissions ? (
-          <DsSkeleton
-            width='100%'
-            height='2.5rem'
+      <DelegatedResourcesSection
+        resourcePermissions={resourcePermissions}
+        isFetching={isFetching}
+        hasError={!!resourcesError}
+        onRemove={handleRemove}
+        isResourceLoading={isLoading}
+        addNewResourceButton={
+          <DelegationModal
+            delegationType={DelegationType.MaskinportenScope}
+            availableActions={[DelegationAction.DELEGATE]}
           />
-        ) : (
-          <ScopeList
-            addNewResourceButton={
-              <DelegationModal
-                delegationType={DelegationType.MaskinportenScope}
-                availableActions={[DelegationAction.DELEGATE]}
-              />
-            }
-            resources={filteredResources}
-            search={search}
-            setSearch={setSearch}
-            filterState={filterState}
-            setFilterState={setFilterState}
-            serviceOwnerOptions={serviceOwnerOptions}
-            onSelect={(resource) => {
-              setSelectedResource(resource);
-              scopeModalRef.current?.showModal();
-            }}
-            renderControls={(resource) => {
-              const resourceLoading = isLoading(resource.identifier);
-              return (
-                <Button
-                  variant='tertiary'
-                  size='sm'
-                  loading={resourceLoading}
-                  aria-disabled={resourceLoading}
-                  onClick={() => {
-                    if (resourceLoading) return;
-                    handleRemove(resource);
-                  }}
-                  aria-label={t('common.delete_poa_for', { poa_object: resource.title })}
-                >
-                  <MinusCircleIcon aria-hidden='true' />
-                  {!isMobile && t('common.delete_poa')}
-                </Button>
-              );
-            }}
-            emptyState={
-              <DsParagraph data-size='md'>
-                {delegatedResources.length === 0
-                  ? t('maskinporten_page.no_delegated_resources')
-                  : t('resource_list.no_resources_filtered', { searchTerm: search })}
-              </DsParagraph>
-            }
-          />
-        )}
-        <EditModal
-          ref={scopeModalRef}
-          maskinportenScope={selectedResource ?? undefined}
-          onClose={() => setSelectedResource(null)}
-        />
-        <MaskinportenDeleteDialog
-          ref={deleteDialogRef}
-          heading={t('maskinporten_page.remove_supplier_heading')}
-          body={t('maskinporten_page.remove_supplier_body', { name: supplierName })}
-          confirmLabel={t('maskinporten_page.remove_supplier_confirm')}
-          onConfirm={handleConfirmDeleteSupplier}
-          isLoading={isRemovingSupplier}
-        />
-      </section>
+        }
+      />
+      <MaskinportenDeleteDialog
+        ref={deleteDialogRef}
+        heading={t('maskinporten_page.remove_supplier_heading')}
+        body={t('maskinporten_page.remove_supplier_body', { name: supplierName })}
+        confirmLabel={t('maskinporten_page.remove_supplier_confirm')}
+        onConfirm={handleConfirmDeleteSupplier}
+        isLoading={isRemovingSupplier}
+      />
     </PageContainer>
   );
 };
